Cleanup javascript and move to json
This commit changes the app.js to use awscors.coffee. This makes
modifying and understanding the javascript logic much easier.

We also add jQuery to the mix of javascript. This allows us to write
simpler scripts and hand over the error handling to proven libraries.

In addition to coffeescript, there is a new dependency json. Instead of
hacking our way through various quirks, we just transfer the url within
json and let those libraries handle the quirks.